requirements.txt
logging
python-telegram-bot==13.7
urllib3==1.26.15
python-dotenv
paramiko
